Move throwToResourceManager to MonadResourceManager
This adds the capability to report exceptions from both ResourceManagerIO and
ResourceManagerSTM, further increasing feature parity.

Adds a timeout to the root resource manager before exceptions are logged
to the console (usually they are thrown as a CombinedException, so
logging is only required for debugging visibility when resource
management is deadlocked).