Remove extra-substituters entry for cachix from flake.nix
The entry cannot be used without nix trusted-user permissions but will
ask for trust / log warnings on every run when running without
trusted-user permissions.

Either this can be reintroduced as a readme section, or we might change
the caching solution to use the magic-nix-cache (which uses the github
action cache) to extend caching for more branches and PRs.