quasar: Store context information for merging deltas in PendingChange
This change adds an additional associated data type to a container which
can be used to store additional information about the container, e.g.
the current length of the list. This is required to efficiently repack
merged list deltas.

Temporarily breaks the `bind`-implementation, since merge context
tracking is integrated into the `PendingChange` abstraction, which is
not yet used by `bind` (bind predates PendingChange).