quasar: Replace RcVar with Rc, with tryExtractRc function
Extracting the value from an RcVar was always planned, but having an
option to not call the cleanup function as part of the RcVar api will
lead to problems in use-cases where that's illegal.

To prevent future issues, the Rc type (previously RcVar) is now
readonly. The cleanup function has been removed, the content is required
to be disposable. This allows to dispose the content after using
`tryExtractRc`.

If required, a mutable variant can be added back later (without the
option to extract the content).